move rules to their own classes
