+++ wayang-api/wayang-api-sql/src/main/java/org/apache/wayang/api/sql/calcite/rules/WayangProjectRule.java
@@ -1,0 +1,53 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to you under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.wayang.api.sql.calcite.rules;
+
+import org.apache.calcite.plan.Convention;
+import org.apache.calcite.rel.RelNode;
+import org.apache.calcite.rel.convert.ConverterRule;
+import org.apache.calcite.rel.logical.LogicalProject;
+
+import org.apache.wayang.api.sql.calcite.convention.WayangConvention;
+import org.apache.wayang.api.sql.calcite.rel.WayangProject;
+
+/**
+ * Rule that converts {@link LogicalProject} to Wayang convention
+ * {@link WayangProject}
+ */
+public class WayangProjectRule extends ConverterRule {
+    public static final Config DEFAULT_CONFIG = Config.INSTANCE
+            .withConversion(LogicalProject.class,
+                    Convention.NONE, WayangConvention.INSTANCE,
+                    "WayangProjectRule")
+            .withRuleFactory(WayangProjectRule::new);
+
+    protected WayangProjectRule(final Config config) {
+        super(config);
+    }
+
+    public RelNode convert(final RelNode rel) {
+        final LogicalProject project = (LogicalProject) rel;
+        return new WayangProject(
+                project.getCluster(),
+                project.getTraitSet().replace(WayangConvention.INSTANCE),
+                convert(project.getInput(), project.getInput().getTraitSet()
+                        .replace(WayangConvention.INSTANCE)),
+                project.getProjects(),
+                project.getRowType());
+    }
+}
